Adds Save using Row Header to Bulk Xlsx Interpreter as well, by way of an intermediate abstract class to hold the common code.
