+++ src/Resources/public/js/pimcore/interpreter/bulk-xlsx.js
@@ -32,6 +32,12 @@ pimcore.plugin.pimcoreDataImporterBundle.configuration.components.interpreter.bu
                         name: this.dataNamePrefix + 'skipFirstRow',
                         value: this.data.hasOwnProperty('skipFirstRow') ? this.data.skipFirstRow : false,
                         inputValue: true
+                    },{ 
+                        xtype: 'checkbox',
+                        fieldLabel: t('plugin_pimcore_datahub_data_importer_configpanel_csv_save_row_header'),
+                        name: this.dataNamePrefix + 'saveHeaderName',
+                        value: this.data.hasOwnProperty('saveHeaderName') ? this.data.saveHeaderName : false,
+                        inputValue: true
                     },{
                         xtype: 'textfield',
                         fieldLabel: t('plugin_pimcore_datahub_data_importer_configpanel_xlsx_sheet'),
